Initial commit
I believe the script will crash if it is used on a day in which one of the teams is scheduled to be playing, as the site will be slightly different. However I cannot do anything about this until one such day rolls around!

Another 'issue' is that the events created only last an hour. I could hit tab a few times and sort this out with a little bit of work, but I'm not that bothered to be honest. I just wanted a quick way to get the fixtures into my phone and as long as I have a kickoff time I've succeeded in what I set out to do.

Maybe at some point I'll make it check for duplicates before creating events, and possibly being able to recognise rescheduled games and react accordingly. However, for now I'm satisfied that it does what I initially wanted it to do.