Matchday tweaks
Found myself an hour before kickoff today (then found a late kickoff when I ran out of time :wink:) to play with the aforementioned 'today' issue.

I believe that I have written something sturdy enough to cope with that, but until March now I won't be able to make absolutely sure. The potential issue that I am forseeing is when 'today' is the first match of the month, and therefore there won't be a 'results' tab in the same month. As I say, I think I've written something that will handle that, but until I actually have those circumstances available to test and play with, I can't be 100% sure.

I'll probably comment it a bit better soon too, haven't really bothered this time around :fearful: